oracle-ebs-gemini-enterprise-framework/Agents/EBS_Master/sub-agents/EBS_SQL_Agent/requirements.txt
mcp
google-cloud-aiplatform
google-adk
opentelemetry-instrumentation-google-genai
